chore(deps): bump allowed AWS provider version
This bumps the allowed version for the aws provider to allow version 6.0.0 and later to be used, too.

Another possibility would be to completely remove the upper version constraint, but the change in this PR has fewer implications than the complete removal.
Therefore, I decided to propose this initially.

I'm happy to completely remove the upper version constraint if you'd prefer that.